Add tiled video rendering
`--video` is now compatible with `--tiled`, and when specified together,
the video frames will be rendered offscreen in tiles (as specified by
`--tile-size`. This makes it possible to render very large videos and/
or very expensive shaders. To render each frame in a single tile,
specify a tile size larger than the minimum dimension of the frame.